+++ src/layout/section/SEALandingLayout/SEALandingLayout.unit.tsx
@@ -1,7 +1,0 @@
-// import React from 'react'
-
-import {} from '@testing-library/react'
-
-// import { SEALandingLayoutProps } from './index'
-
-describe('SEALandingLayoutProps', () => {})