+++ src/layout/section/SEALandingLayout/index.tsx
@@ -1,2 +1,0 @@
-export { SEALandingLayout, SEALandingLayoutProps } from './SEALandingLayout'
-export { SEALandingLayoutElements } from './SEALandingLayout.style'