+++ src/layout/section/seaSection/index.tsx
@@ -1,0 +1,1 @@
+export { SEASection, SEASectionProps } from './SeaSection'