+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Unreleased
 
+- **[NEW]** `SEASection` holds SEA Page layout
+
 [...]
 
 # v45.2.0 (21/01/2021)
+
 - **[FIX]** Fix `MediaSizeProvider` first render state
 - **[NEW]** New serverSideMediaSize attribute for `MediaSizeProvider`